Implement delta-time measurements for Engine and change delta_time type for MasterControllers to f64

+++ src/core/world/mc.rs
@@ -2,8 +2,8 @@ use super::World;
 
 pub trait MasterController {
     type ObserverEvent;
-    fn start(&mut self, _world: &mut World, _delta_time: f32) {}
-    fn tick(&mut self, _world: &mut World, _delta_time: f32) -> EngineInstruction { EngineInstruction::Run {
+    fn start(&mut self, _world: &mut World, _delta_time: f64) {}
+    fn tick(&mut self, _world: &mut World, _delta_time: f64) -> EngineInstruction { EngineInstruction::Run {
         run_dispatcher: true
     } }
 }
